requirements.txt
amqp==2.2.2
anyjson==0.3.3
appdirs==1.4.3
celery==4.1.0
Django==1.11
django-filter==1.0.2
django-nose==1.4.4
djangorestframework==3.6.2
packaging==16.8
pyparsing==2.2.0
pytz==2017.2
six==1.10.0
psycopg2==2.7.3.2
pika==0.11.2